feat: adiciona deleção de autores, tratamento de erros e testes

+++ libraryapi/src/main/java/https/github/com/GabrielPCdS/libraryapi/service/AutorService.java
@@ -1,7 +1,11 @@
 package https.github.com.GabrielPCdS.libraryapi.service;
 
+import https.github.com.GabrielPCdS.libraryapi.exceptions.OperacaoNaoPermitidaException;
 import https.github.com.GabrielPCdS.libraryapi.model.Autor;
 import https.github.com.GabrielPCdS.libraryapi.repository.AutorRepository;
+import https.github.com.GabrielPCdS.libraryapi.repository.LivroRepository;
+import https.github.com.GabrielPCdS.libraryapi.validator.AutorValidator;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -9,15 +13,16 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Service
+@RequiredArgsConstructor
 public class AutorService {
 
     private final AutorRepository repository;
+    private final AutorValidator validator;
+    private final LivroRepository livroRepository;
 
-    public AutorService(AutorRepository repository){
-        this.repository = repository;
-    }
 
     public Autor salvar(Autor autor){
+        validator.validar(autor);
         return repository.save(autor);
     }
 
@@ -26,6 +31,7 @@ public class AutorService {
             throw new IllegalArgumentException("Para Atualizar é necessario que o autor já " +
                     "esteja salvo na base.");
         }
+        validator.validar(autor);
         repository.save(autor);
     }
 
@@ -34,9 +40,14 @@ public class AutorService {
     }
 
     public void deletar(Autor autor){
+        if (possuiLivro(autor)){
+            throw new OperacaoNaoPermitidaException(
+                    "Não é permitido excluir um Autor que possui Livros cadastrados!");
+        }
         repository.delete(autor);
-    }
 
+
+    }
     public List<Autor> pesquisa(String name, String nacionalidade){
         if (name != null && nacionalidade != null){
             return repository.findByNameAndNacionalidade(name, nacionalidade);
@@ -49,6 +60,10 @@ public class AutorService {
          return repository.findByNacionalidade(nacionalidade);
      }
      return repository.findAll();
+    }
+
+    public boolean possuiLivro(Autor autor){
+        return livroRepository.existsByAutor(autor);
     }
 
 }
